Add back and submit button, notifications, and more
Hooked the back button up to the back method. Make the back method have
a confirmation popup (using ConfirmationPop() from Kegan's code), hooked
submit button up to the submit method and made it properly send stuff
to the database.

Made a front-end validation method, as-per Kegan's example, to ensure
that everything is filled out in the completedAudit before submission.

Made a clear_page() method, also as_per Kegan's example, which sets the
page back to a default state so nothing gets screwed up if the user
goes back to the page.