Change button and scrollview spacing
While group programming peoples mentioned wanting to alter the spacing
to be more aesthetic, hopefully it is a bit better now. Some issues
are gonna be a more difficult fix, like the way a scrollview enables
sort-of "over-scrolling" which can leave a giant space inbetween
the title and the rest of the audit when scrolling all the way down.