Implement lockout feature.
Added "locked" attribute to auditTemplates. This determines whether a
template has already been used to submit a completedAudit. Defaults
to zero, so it doesn't NEED to be included when building with
auditTemplateBuilder. But I also added the with_lock() method to the
auditTemplateBuilder, in case there you want an auditTemplate to be
locked right away after it is sent to the database, for whatever reason.

A locked template appears greyed out, to distinguish that it is locked.
If the user tries to open a locked template, a popup informs them
that the audit has already been completed.

The admin page has been updated with a "Clear all audit locks" button
which, as the name suggests, unlocks all audit template.

At the moment, we have repeat auditTemplate titles in the database. This
won't, or shouldn't, be the case in the end product as we have decided.
I made the choice to have the method that updates the database objects
and "locks" them affect all objects with the same title, so that all
of these repeats are locked when one is used. This doesn't affect any
bit for the way it will work in the final version of our program.