lots of changes. will need to refactor josiah_modules' name ^_^
josiah_module.kv is the module that gets added to the stack_list
when you add more stuff for the module add it there.
remember to make things you want to update Objectproperties/String properties (which is what I think was catching us up yesterday.)

josiah_module.kv Line 15: this is that clever trick that was used to left justify the text.